Fix homework 4
