--- scrapper/src/test/java/edu/java/scrapper/IntegrationTest.java
@@ -1,52 +1,0 @@
-package edu.java.scrapper;
-
-import java.io.OutputStreamWriter;
-import java.nio.file.Path;
-import liquibase.Liquibase;
-import liquibase.database.Database;
-import liquibase.database.DatabaseFactory;
-import liquibase.database.jvm.JdbcConnection;
-import liquibase.resource.DirectoryResourceAccessor;
-import lombok.SneakyThrows;
-import lombok.extern.log4j.Log4j2;
-import org.springframework.test.context.DynamicPropertyRegistry;
-import org.springframework.test.context.DynamicPropertySource;
-import org.testcontainers.containers.JdbcDatabaseContainer;
-import org.testcontainers.containers.PostgreSQLContainer;
-import org.testcontainers.junit.jupiter.Testcontainers;
-
-@Testcontainers
-@Log4j2
-public abstract class IntegrationTest {
-    public static PostgreSQLContainer<?> POSTGRES;
-
-    static {
-        POSTGRES = new PostgreSQLContainer<>("postgres:15")
-            .withDatabaseName("scrapper")
-            .withUsername("postgres")
-            .withPassword("postgres");
-        POSTGRES.start();
-
-        runMigrations(POSTGRES);
-    }
-
-    @SneakyThrows
-    private static void runMigrations(JdbcDatabaseContainer<?> container) {
-        Database database = DatabaseFactory.getInstance()
-            .findCorrectDatabaseImplementation(new JdbcConnection(container.createConnection("")));
-        Liquibase liquibase = new Liquibase(
-            "master.xml",
-            new DirectoryResourceAccessor(Path.of("migrations")),
-            database
-        );
-        liquibase.update("", new OutputStreamWriter(System.out));
-
-    }
-
-    @DynamicPropertySource
-    static void jdbcProperties(DynamicPropertyRegistry registry) {
-        registry.add("spring.datasource.url", POSTGRES::getJdbcUrl);
-        registry.add("spring.datasource.username", POSTGRES::getUsername);
-        registry.add("spring.datasource.password", POSTGRES::getPassword);
-    }
-}
